got chats working!
I basically made a "messages" global array variable in Vuex and
whenever you go to a new chat, it runs bindToChat which just
binds right over the "messages" variable with the messages from
that particular chat. It's kinda amazing how this works.

I'm not sure if this is the best way to do this? I didn't want to just
naively store all the messages of every chat in memory because it's
less efficient and I just wanted to see how easy this could be.

And it really is pretty easy. I could even unbind on leaving the chat
view if I wanted to save memory that way. However, it's already
unusually slow when I serve it locally and switch routes so idk if
it's a good idea to have more code running on switch. Maybe this
problem will go away when I build it and figure out how to deploy!